src/core: Changes EventId from uint32_t to uint64_t

+++ src/core/model/default-simulator-impl.h
@@ -117,9 +117,9 @@ private:
   Ptr<Scheduler> m_events;
 
   /** Next event unique id. */
-  uint32_t m_uid;
+  uint64_t m_uid;
   /** Unique id of the current event. */
-  uint32_t m_currentUid;
+  uint64_t m_currentUid;
   /** Timestamp of the current event. */
   uint64_t m_currentTs;
   /** Execution context of the current event. */
